Fixes after doing some load-testing with +100K jobs: - fix timer query, should take in account the lock owner, otherwise two or more nodes can at the same time fetch the same jobs and lock them - between acquiring and moving the timer job to the executable job table, there can be a small window where the job could be deleted (eg timer boundary event on user task), so added extra checks for related execution - The timer acquire and the reset expired queries did not yet apply paging

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/asyncexecutor/DefaultJobManager.java
@@ -105,30 +105,35 @@ public class DefaultJobManager implements JobManager {
     
     CommandContext commandContext = Context.getCommandContext();
     JobEntity executableJob = createExecutableJobFromOtherJob(timerJob, commandContext);
-    commandContext.getJobEntityManager().insert(executableJob);
-    commandContext.getTimerJobEntityManager().delete(timerJob);
+    boolean insertSuccesful = commandContext.getJobEntityManager().insertJobEntity(executableJob);
+    if (insertSuccesful) {
+      commandContext.getTimerJobEntityManager().delete(timerJob);
+      
+      // When the async executor is activated, the job is directly passed on to the async executor thread
+      if (isAsyncExecutorActive()) {
+        hintAsyncExecutor(executableJob); 
+      }
     
-    // When the async executor is activated, the job is directly passed on to the async executor thread
-    if (isAsyncExecutorActive()) {
-      hintAsyncExecutor(executableJob); 
+      return executableJob;
     }
-    
-    return executableJob;
+    return null;
   }
   
   @Override
   public TimerJobEntity moveJobToTimerJob(AbstractJobEntity job) {
     CommandContext commandContext = Context.getCommandContext();
     TimerJobEntity timerJob = createTimerJobFromOtherJob(job, commandContext);
-    commandContext.getTimerJobEntityManager().insert(timerJob);
-    
-    if (job instanceof JobEntity) {
-      commandContext.getJobEntityManager().delete((JobEntity) job);
-    } else if (job instanceof SuspendedJobEntity) {
-      commandContext.getSuspendedJobEntityManager().delete((SuspendedJobEntity) job);
+    boolean insertSuccesful = commandContext.getTimerJobEntityManager().insertTimerJobEntity(timerJob);
+    if (insertSuccesful) {
+      if (job instanceof JobEntity) {
+        commandContext.getJobEntityManager().delete((JobEntity) job);
+      } else if (job instanceof SuspendedJobEntity) {
+        commandContext.getSuspendedJobEntityManager().delete((SuspendedJobEntity) job);
+      }
+      
+      return timerJob;
     }
-    
-    return timerJob;
+    return null;
   }
   
   @Override
